Worked on Assignment2 jptr notebook
Created the helper functions
- dataset loader
- train / test splitting
- 3 model piplelines (logistic regression, multinomialNB, Kmeans)
- 3 training / Evaluation of each model

Prints all the results, plus a dataset overview.

All that needs doing is Data visualisation and In depth descriptions of functions, results etc.

Still waiiting for the updated data, so modifications will need to be made in regards to processing data / pipelines to accomodate new features

👽 👍